base entity + secure msg filter fix

+++ src/main/java/sample/repository/MessageReceiveHistoryRepository.java
@@ -6,5 +6,5 @@ import sample.model.MessageReceiveHistory;
 import sample.model.Subscription;
 
 @Repository
-public interface MessageReceiveHistoryRepository extends JpaRepository<MessageReceiveHistory, Integer> {
+public interface MessageReceiveHistoryRepository extends JpaRepository<MessageReceiveHistory, Long> {
 }
